+++ src/restfw/__init__.py
@@ -24,6 +24,10 @@ def includeme(config):
     config.add_directive('add_sub_resource_fabric', add_sub_resource_fabric)
     config.add_directive('add_sub_resource_fabric_predicate', add_sub_resource_fabric_predicate)
 
+    from .config import add_external_link_fabric, add_external_link_fabric_predicate
+    config.add_directive('add_external_link_fabric', add_external_link_fabric)
+    config.add_directive('add_external_link_fabric_predicate', add_external_link_fabric_predicate)
+
     from .usage_examples.config import add_usage_examples_fabric_predicate
     from .usage_examples.config import add_usage_examples_fabric
     config.add_directive('add_usage_examples_fabric', add_usage_examples_fabric)